feat(proposal): add per-section copy SOW as plaintext button
Extract Tiptap-to-plaintext conversion from DocuSign service into
shared utility (tiptap-to-text.ts) so it can be reused client-side.
Add a copy button to each SOW accordion section that converts that
section's Tiptap JSON to clean plaintext and copies to clipboard.

Closes #35
